Update VirtualGPIO.py
Inserted a full lot of stuff:
VirtualGPIOBaseHandler for handling signals. It should be subclassed
BaseSignal to create Signals.
Creating a signal is as easy as ConnectSignal = BaseSignal("Connect"), because BaseSignal is a MetaClass. 
You, then, can instantiate the signal with the arguments as ConnectSignal(uuid, "20161018001800"), p.e.
Good night you folks!

Tomorrow I'll test all of this. I'll be surprised if this works!
TODO: All the doc stuff... You know.